Made cars reroll on completion
Changed car behavior at the end of the road. Instead of continuing as usual, they reroll their attributes.

This should ensure that lon simulation data is representative of the actual "asocial rate." i.e. otherwise it may be possible for the 0.1 asocial case to have 0 asocial cars for low car counts.